==> python/1-abebooks.ipynb <==
# Abebooks

___

import pandas as pd
import os
import glob
import mysql.connector

os.chdir("C:\Adrian - GDrive\Professionnel\Formation\Informatique - Digital\Data University\Projet certification\src\\abebooks")

df = pd.DataFrame(columns=[], index=[])

for file in glob.glob("*.txt"):
    #print(file)
    csv = pd.read_csv(file, sep="\t")
    df = df.append(csv, ignore_index=True)

df.shape

df.columns

df = df.drop(columns=['N° de commande Abebooks', 'Programme', "Option d'achat",
       'Auteur', 'Titre',
       'Frais de port (EUR)', 'Paiements pour vos ventes (brut) (EUR)',
       'Commissions Abebooks (EUR)',
       'Frais de prestataire de paiement externe (EUR)',
       'Frais de prestataire de paiement externe (pour cent)',
       'Paiements pour vos ventes (net) (EUR)',
       'Date du retour', 'Unnamed: 15'])


df.shape

df.head()

df.dtypes

df['Date de traitement'] = pd.to_datetime(df['Date de traitement'], format='%d/%m/%Y')

df = df[df["Prix du livre (EUR)"] > "0"]
df

df["Prix du livre (EUR)"] = df["Prix du livre (EUR)"].str.replace(",",".").str.replace(" ", "").str.replace("\xa0", "").astype(float)

df.sort_values(by="Prix du livre (EUR)")

os.chdir("C:\Adrian - GDrive\Professionnel\Formation\Informatique - Digital\Data University\Projet certification\csv")

df.to_csv("1-abebooks.csv", index=False)
