==> python/1-amazon.ipynb <==
# Amazon

___

import pandas as pd
import os
import glob

os.chdir("C:\Adrian - GDrive\Professionnel\Formation\Informatique - Digital\Data University\Projet certification\src\\amazon")

df = pd.DataFrame(columns=[], index=[])

for file in glob.glob("*.txt"):
    #print(file)
    csv = pd.read_csv(file, sep="\t")
    df = df.append(csv, ignore_index=True)

df.shape

df.columns

df = df.drop(columns=['order-id', 'order-item-id', 'payments-date', 'buyer-email', 'buyer-name', 'buyer-phone-number', 'product-name', 'quantity-purchased', 'currency', 'item-tax', 'shipping-price', 'shipping-tax', 'ship-service-level', 'recipient-name', 'ship-address-1', 'ship-address-2', 'ship-address-3', 'ship-city', 'ship-state', 'ship-postal-code', 'ship-phone-number', 'bill-address-1', 'bill-address-2', 'bill-address-3', 'bill-city', 'bill-state', 'bill-postal-code','bill-country', 'delivery-start-date', 'delivery-end-date',       'delivery-time-zone', 'delivery-Instructions', 'sales-channel', 'bill-name'])


df.shape

df.dtypes

df['purchase-date'] = pd.to_datetime(df['purchase-date'].astype(str).str.slice(0,10), format='%Y-%m-%d')
df.head()

os.chdir("C:\Adrian - GDrive\Professionnel\Formation\Informatique - Digital\Data University\Projet certification\src")
df_country = pd.read_csv("sql-pays.csv")
df_country

df = df.merge(df_country, left_on='ship-country', right_on='Alpha2')
df.columns

df = df.drop(columns=['ship-country', 'Id', 'Code', 'Alpha2', 'Alpha3', 'nom_fr_fr'])
df

os.chdir("C:\Adrian - GDrive\Professionnel\Formation\Informatique - Digital\Data University\Projet certification\csv")

df.to_csv("1-amazon.csv", index=False)
