==> python/1-ebay.ipynb <==
# eBay

---

import pandas as pd
import os
import glob

os.chdir("C:\Adrian - GDrive\Professionnel\Formation\Informatique - Digital\Data University\Projet certification\src\\ebay")

df_raw = pd.DataFrame(columns=[], index=[])

for file in glob.glob("*.csv"):
    #print(file)
    csv = pd.read_csv(file, sep=";", encoding='latin-1')
    df_raw = df_raw.append(csv, ignore_index=True)

df_raw.shape


df = df_raw

df.columns

df = df.drop(columns=(["Numéro de la fiche de vente", "Pseudo", "Nom complet de l'acheteur", "E-mail de l'acheteur", "Acheteur - Adresse 1", "Acheteur - Adresse 2", "Ville", "Région", "Code postal", "Pays", "Code de la commande", "Numéro de l'objet", "Code de la transaction ", "Titre de l'objet", "Quantité", "Taux de TVA inclus", "Livraison et expédition", "Assurance", "Taxe perçue par eBay", "Prix total", "Mode de paiement", "Code de transaction PayPal", "Numéro de la facture", "Date de facturation", "Date de la finalisation de l'achat", "Payé le", "Envoyé le", "Service de livraison", "Évaluation laissée", "Évaluation reçue", "Mes commentaires", "Commentaires", "Type de code produit", "Valeur de code produit", "Product ID Value 2", "Détails des différentes versions", "Code de référence produit", "Numéro de suivi ", "Code de référence pour la livraison internationale", "Adresse de livraison 1", "Adresse de livraison 2", "Ville de livraison", "Région de livraison", "Code postal de livraison", "Téléphone"]))

df.head()

df.dtypes

df["Prix de vente"] = df["Prix de vente"].str.replace(",",".").str.replace(" ","").str.slice(0,-4).astype(float)
df.describe()

df['Date de vente'] = df['Date de vente'].str.replace("déc.","12").str.replace("janv.","01").str.replace("févr.","02").str.replace("mars","03").str.replace("avr.","04").str.replace("mai","05").str.replace("juin","06").str.replace("juil.","07").str.replace("août","08")
df['Date de vente'] = pd.to_datetime(df['Date de vente'], format="%d-%m-%y")
df

df["Pays de livraison"] = df["Pays de livraison"].str.replace("France métropolitaine","France")
df

os.chdir("C:\Adrian - GDrive\Professionnel\Formation\Informatique - Digital\Data University\Projet certification\src")
df_country = pd.read_csv("sql-pays.csv")
df_country

df = df.merge(df_country, left_on='Pays de livraison', right_on='nom_fr_fr')
df.columns

df = df.drop(columns=['Pays de livraison', 'Id', 'Code', 'Alpha2', 'Alpha3', 'nom_fr_fr'])
df

df = df[df["Libellé personnalisé"].notna()]

os.chdir("C:\Adrian - GDrive\Professionnel\Formation\Informatique - Digital\Data University\Projet certification\csv")

df.to_csv("1-ebay.csv", index=False)
