==> python/3-sales.ipynb <==
# A/ Concaténation des fichiers de ventes

import pandas as pd
import numpy as np
import os
import mysql.connector

os.chdir("C:\Adrian - GDrive\Professionnel\Formation\Informatique - Digital\Data University\Projet certification\csv")

df_abebooks = pd.read_csv("1-abebooks.csv")
df_amazon = pd.read_csv("1-amazon.csv")
df_ebay = pd.read_csv("1-ebay.csv")
df_fnac = pd.read_csv("1-fnac.csv")
df_lelivre = pd.read_csv("1-lelivre.csv")
df_leslibraires = pd.read_csv("1-leslibraires.csv")
df_rakuten = pd.read_csv("1-rakuten.csv")

df_abebooks["sale_source"] = "abebooks" 
df_amazon["sale_source"] = "amazon" 
df_ebay["sale_source"] = "ebay" 
df_fnac["sale_source"] = "fnac"
df_lelivre["sale_source"] = "lelivre"  
df_leslibraires["sale_source"] = "leslibraires" 
df_rakuten["sale_source"] = "rakuten" 

print(df_abebooks.columns)
print(df_amazon.columns)
print(df_ebay.columns)
print(df_fnac.columns)
print(df_lelivre.columns)
print(df_leslibraires.columns)
print(df_rakuten.columns)

df_abebooks = df_abebooks.rename(columns={'Votre n° de référence': 'sku', 'Prix du livre (EUR)': 'item_price', 'Date de traitement': 'purchase_date'})
df_amazon = df_amazon.rename(columns={'purchase-date': 'purchase_date', 'nom_en_gb': 'ship_country', 'item-price': 'item_price'})
df_ebay = df_ebay.rename(columns={'Date de vente': 'purchase_date', 'nom_en_gb': 'ship_country', 'Prix de vente': 'item_price', 'Libellé personnalisé': 'sku'})
df_fnac = df_fnac.rename(columns={'Date de création': 'purchase_date', 'nom_en_gb': 'ship_country', 'Prix total (TTC)': 'item_price', 'SKUs': 'sku'})
df_lelivre = df_lelivre.rename(columns={'nom_en_gb': 'ship_country'})
df_leslibraires = df_leslibraires.rename(columns={'Date de commande': 'purchase_date', 'Total TTC': 'item_price'})
df_rakuten = df_rakuten.rename(columns={'Date commande': 'purchase_date', 'Mandat article EUR': 'item_price', "Réf. Vendeur":"sku"})


print(df_abebooks.shape)
print(df_amazon.shape)
print(df_ebay.shape)
print(df_fnac.shape)
print(df_lelivre.shape)
print(df_leslibraires.shape)
print(df_rakuten.shape)

print(df_abebooks.head())
print(df_amazon.head())
print(df_ebay.head())
print(df_fnac.head())
print(df_lelivre.head())
print(df_leslibraires.head())
print(df_rakuten.head())

df_full_raw = pd.concat([df_abebooks,df_amazon,df_ebay,df_fnac,df_lelivre,df_leslibraires,df_rakuten])

df_full_raw.dtypes

df_full_raw.shape

df_full_raw.head()

df_full_raw.describe(include='all')

# Choix de la période
df_full_raw = df_full_raw[df_full_raw["purchase_date"] > "2018-12-31"]
df_full_raw = df_full_raw[df_full_raw["purchase_date"] < "2019-09-01"]
# Suppression des remboursement
df_full_raw = df_full_raw[df_full_raw["item_price"] > 0]

df_full_raw.shape

df_full_raw.groupby("sale_source").count()

----

# B/ Ajout d'informations complémentaires

## Connexion à la base locale pour réupérér le dewey, creation date et printing year

try:
    connection = mysql.connector.connect(host='localhost',
                                         database='db_lelivre_pro',
                                         user='localscript',
                                         password='GY9x9xLyaJA8CaqQMG6S',
                                         port='3308')
    if connection.is_connected():
        db_Info = connection.get_server_info()
        print("Connected to MySQL Server version ", db_Info)
        cursor = connection.cursor()
        cursor.execute("select database();")
        record = cursor.fetchone()
        print("You're connected to database: ", record)
except Error as e:
    print("Error while connecting to MySQL", e)
finally:
    if (connection.is_connected()):
        ### Dataframe 1
        var_sql = "SELECT ref as sku, creationdate as creation_date, type as dewey, printingyear as printing_year, ean, isbn, titre FROM db_lelivre_pro.livresv2"
        df_info = pd.read_sql(var_sql, connection)
        
        print("Dataframe constitué")
        cursor.close()
        connection.close()
        print("MySQL connection is closed")

df_full = df_full_raw.merge(df_info, on="sku")
df_full.shape

df_full

---

# C/ Transformation des variables

df_full.dtypes

## Epoque

df_full["printing_year"] = df_full["printing_year"].astype(int)
# Suppression des dates incohérentes
df_full = df_full[df_full["printing_year"] > 1600]
df_full = df_full[df_full["printing_year"] < 2020]
df_full

df_full["century"] = 0
df_full["century"][df_full["printing_year"].astype(str).str.slice(0,2) == "16"] = "XVIIe"
df_full["century"][df_full["printing_year"].astype(str).str.slice(0,2) == "17"] = "XVIIIe"
df_full["century"][df_full["printing_year"].astype(str).str.slice(0,2) == "18"] = "XIXe"
df_full["century"][df_full["printing_year"].astype(str).str.slice(0,2) == "19"] = "XXe"
df_full["century"][df_full["printing_year"].astype(str).str.slice(0,2) == "20"] = "XXIe"
df_full

## Thématique

df_full = df_full[df_full["dewey"] < 1000]
df_full["dewey_class"] = ((df_full["dewey"] / 100).apply(np.floor) * 100).astype(int).astype(str)
df_full["dewey"] = df_full["dewey"].astype(str)

df_dewey = pd.read_csv("C:\Adrian - GDrive\Professionnel\Formation\Informatique - Digital\Data University\Projet certification\src\dewey_class.csv")
df_dewey["dewey_class"] = df_dewey["dewey_class"].astype(str)
df_dewey.dtypes

df_full = df_full.merge(df_dewey)

## Délai de rétention

df_full["creation_date"] = pd.to_datetime(df_full["creation_date"].astype(str).str.slice(0,8), format="%Y%m%d")
df_full["purchase_date"] = pd.to_datetime(df_full["purchase_date"], format="%Y-%m-%d")

df_full["retention_delay_years"] = ((df_full["purchase_date"] - df_full["creation_date"]) / np.timedelta64(1,'Y'))
df_full = df_full[df_full["retention_delay_years"] > 0]

## Expédition : Région du monde

os.chdir("C:\Adrian - GDrive\Professionnel\Formation\Informatique - Digital\Data University\Projet certification\src")
df_region = pd.read_csv("countries of the world.csv")

df_region["Country"] = df_region["Country"].str.slice(0,-1)

df_full = df_full.merge(df_region, left_on='ship_country', right_on='Country', how='left')
df_full.shape

df_full = df_full.rename(columns={'Region': 'ship_world_region'})

df_full["ship_world_region"] = df_full["ship_world_region"].str.strip()
df_full["ship_world_region"] = df_full["ship_world_region"].str.replace(" ","_")

df_full["ship_world_region"][(df_full["ship_world_region"] != "WESTERN_EUROPE") & (df_full["ship_world_region"] != "NORTHERN_AMERICA") & ~df_full["ship_world_region"].isna()] = "REST_OF_THE_WORLD"

df_full["ship_world_region"][(df_full["ship_country"] == "France")] = "FRANCE"

df_full.describe(include='all')

## Nettoyage du dataset

df_full = df_full.drop(columns=["purchase_date", "creation_date", "dewey", 'Country', 'Population', 'Area (sq. mi.)', 'Pop. Density (per sq. mi.)', 'Coastline (coast/area ratio)', 'Net migration', 'Infant mortality (per 1000 births)', 'GDP ($ per capita)', 'Literacy (%)', 'Phones (per 1000)', 'Arable (%)', 'Crops (%)', 'Other (%)', 'Climate', 'Birthrate', 'Deathrate', 'Agriculture', 'Industry', 'Service'])
df_full

## Exportation

os.chdir("C:\Adrian - GDrive\Professionnel\Formation\Informatique - Digital\Data University\Projet certification\csv")
df_full.to_csv("3-sales.csv", index=False)

---

# Statistiques descriptives

df_full.describe()

df_full.sort_values(by="item_price", ascending=False).head(10)

df_full.sort_values(by="item_price", ascending=False)

df_full[["sku","century"]].groupby("century").count().sort_values(by="sku", ascending=False)

df_full["isbn"][df_full["isbn"] != ""].count()

df_full["ean"][df_full["ean"] != ""].count()

df_full[(df_full["ean"] != "") & (df_full["isbn"] != "")]
