==> python/5-dataset.ipynb <==
# Merge des fichiers sales et catalog

import pandas as pd
import os

os.chdir("C:\Adrian - GDrive\Professionnel\Formation\Informatique - Digital\Data University\Projet certification\csv")

df_sales = pd.read_csv("3-sales.csv")
df_catalog_isbn = pd.read_csv("4-catalog_rakuten_isbn.csv")
df_catalog_ean = pd.read_csv("4-catalog_rakuten_ean.csv")

df_catalog_ean.shape

df_catalog_isbn.shape

df_catalog_ean[["ean","value"]].groupby("value").count().sort_values(by="ean", ascending=False).head(1)

df_catalog_isbn[["isbn","value"]].groupby("value").count().sort_values(by="isbn", ascending=False).head(1)

df_catalog = df_catalog_ean.merge(df_catalog_isbn, left_index=True, right_index=True)

df_catalog.shape

df_catalog["value_rakuten"] = "None"
df_catalog["quantity_rakuten"] = "None"

df_catalog["value_rakuten"][df_catalog["value_y"] != "None"] = df_catalog["value_y"]
df_catalog["quantity_rakuten"][df_catalog["value_y"] != "None"] = df_catalog["quantity_y"]

df_catalog["value_rakuten"][df_catalog["value_x"] != "None"] = df_catalog["value_x"]
df_catalog["quantity_rakuten"][df_catalog["value_x"] != "None"] = df_catalog["quantity_x"]

df_catalog.head(40)

df_catalog[["isbn","value_rakuten"]].groupby("value_rakuten").count().sort_values(by="isbn", ascending=False).head(1)

df_catalog = df_catalog.drop(columns=["value_x","quantity_x","value_y","quantity_y"])
df_catalog

---

df_sales["ean"] = df_sales["ean"].astype(str).str.slice(0,-2)

df_sales.dtypes

df_tmp = df_sales[df_sales["isbn"].notna()].reset_index().merge(df_catalog, left_index=True, right_index=True)

df_tmp.columns

df_tmp = df_tmp.drop(columns=(['sku', 'item_price', 'sale_source', 'ship_country',
       'printing_year', 'titre',  'ean_x', 'isbn_x', 'century', 'dewey_class',
       'retention_delay_years', 'ship_world_region', 'ean_y', 'isbn_y']))

df = df_sales.merge(df_tmp, left_index=True, right_on="index", how='left')

df

---

df.sort_values(by="item_price", ascending=False)

df["url_photo"] = "https://www.le-livre.fr/photos/" + df["sku"].str.slice(0,3) + "/" + df["sku"] + ".jpg"

df.to_csv("5-dataset.csv", index=False)

---

# Analyse

df.dtypes
